+++ react/Drawer.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState, useEffect, ReactElement } from 'react'
+import React, { ReactElement, Suspense, useReducer } from 'react'
 import { defineMessages } from 'react-intl'
 
 import { IconClose, IconMenu } from 'vtex.store-icons'
@@ -6,112 +6,59 @@ import { useCssHandles } from 'vtex.css-handles'
 
 import Overlay from './Overlay'
 import Portal from './Portal'
-import Swipable from './Swipable'
+import useLockScroll from './modules/useLockScroll'
 
-// https://stackoverflow.com/a/3464890/5313009
-const getScrollPosition = () => {
-  const documentElement =
-    window && window.document && window.document.documentElement
-  if (!documentElement) {
-    return 0
-  }
-  return (
-    (window.pageYOffset || documentElement.scrollTop) -
-    (documentElement.clientTop || 0)
-  )
+const Swipable = React.lazy(() => import('./Swipable'))
+
+interface MenuState {
+  isOpen: boolean
+  hasBeenOpened: boolean
 }
 
-const useLockScroll = () => {
-  const [isLocked, setLocked] = useState(false)
-  type ScrollPosition = number | null
-  const [lockedScrollPosition, setLockedScrollPosition] = useState<
-    ScrollPosition
-  >(null)
+interface MenuAction {
+  type: 'open' | 'close'
+}
 
-  useEffect(() => {
-    /** Locks scroll of the root HTML element if the
-     * drawer menu is open
-     */
-    const shouldLockScroll = isLocked
+const initialMenuState: MenuState = {
+  isOpen: false,
+  hasBeenOpened: false,
+}
 
-    const documentElement =
-      window && window.document && window.document.documentElement
-    if (documentElement) {
-      documentElement.style.overflow = shouldLockScroll ? 'hidden' : 'auto'
-
-      /** iOS doesn't lock the scroll of the body by just setting overflow to hidden.
-       * It requires setting the position of the HTML element to fixed, which also
-       * resets the scroll position.
-       * This code is intended to record the scroll position and set it as
-       * the element's position, and revert it once the menu is closed.
-       */
-      const scrollPosition =
-        lockedScrollPosition == null
-          ? getScrollPosition()
-          : lockedScrollPosition
-
-      if (lockedScrollPosition == null && shouldLockScroll) {
-        setLockedScrollPosition(scrollPosition)
+function menuReducer(state: MenuState, action: MenuAction) {
+  switch (action.type) {
+    case 'open':
+      return {
+        ...state,
+        isOpen: true,
+        hasBeenOpened: true,
       }
-
-      if (lockedScrollPosition != null && !shouldLockScroll) {
-        window && window.scrollTo(0, scrollPosition)
-        setLockedScrollPosition(null)
+    case 'close':
+      return {
+        ...state,
+        isOpen: false,
       }
-
-      documentElement.style.position = shouldLockScroll ? 'fixed' : 'static'
-
-      documentElement.style.top = shouldLockScroll
-        ? `-${scrollPosition}px`
-        : 'auto'
-
-      documentElement.style.bottom = shouldLockScroll ? '0' : 'auto'
-      documentElement.style.left = shouldLockScroll ? '0' : 'auto'
-      documentElement.style.right = shouldLockScroll ? '0' : 'auto'
-    }
-
-    return () => {
-      documentElement.style.overflow = 'auto'
-      documentElement.style.position = 'static'
-
-      documentElement.style.top = 'auto'
-      documentElement.style.bottom = 'auto'
-      documentElement.style.left = 'auto'
-      documentElement.style.right = 'auto'
-    }
-  }, [isLocked]) // eslint-disable-line react-hooks/exhaustive-deps
-  // ☝️ no need to trigger this on lockedScrollPosition changes
-
-  return setLocked
+    default:
+      return state
+  }
 }
 
 const useMenuState = () => {
-  const [isMenuOpen, setIsOpen] = useState(false)
-  const [isMenuTransitioning, setIsTransitioning] = useState(false)
+  const [state, dispatch] = useReducer(menuReducer, initialMenuState)
   const setLockScroll = useLockScroll()
 
-  let transitioningTimeout: number | null
-
   const setMenuOpen = (value: boolean) => {
-    setIsOpen(value)
-    setIsTransitioning(true)
+    dispatch({ type: value ? 'open' : 'close' })
     setLockScroll(value)
-
-    if (transitioningTimeout != null) {
-      clearTimeout(transitioningTimeout)
-      transitioningTimeout = null
-    }
-    transitioningTimeout =
-      window &&
-      window.setTimeout(() => {
-        setIsTransitioning(false)
-      }, 300)
   }
 
   const openMenu = () => setMenuOpen(true)
   const closeMenu = () => setMenuOpen(false)
 
-  return { isMenuOpen, isMenuTransitioning, setMenuOpen, openMenu, closeMenu }
+  return {
+    state,
+    openMenu,
+    closeMenu,
+  }
 }
 
 const CSS_HANDLES = [
@@ -125,10 +72,6 @@ const CSS_HANDLES = [
 const Drawer: StorefrontComponent<
   DrawerSchema & { customIcon: ReactElement }
 > = ({
-  // actionIconId,
-  // dismissIconId,
-  // position,
-  // height,
   width,
   customIcon,
   maxWidth = 450,
@@ -136,37 +79,16 @@ const Drawer: StorefrontComponent<
   slideDirection = 'horizontal',
   children,
 }) => {
-  const {
-    isMenuOpen,
-    isMenuTransitioning,
-    openMenu,
-    closeMenu,
-  } = useMenuState()
+  const { state: menuState, openMenu, closeMenu } = useMenuState()
+  const { isOpen: isMenuOpen, hasBeenOpened: hasMenuBeenOpened } = menuState
   const handles = useCssHandles(CSS_HANDLES)
-  const menuRef = useRef(null)
 
-  const slideFromTopToBottom = `translate3d(0, ${
-    isMenuOpen ? '0' : '-100%'
-  }, 0)`
-  const slideFromLeftToRight = `translate3d(${
-    isMenuOpen ? '0' : '-100%'
-  }, 0, 0)`
-  const slideFromRightToLeft = `translate3d(${isMenuOpen ? '0' : '100%'}, 0, 0)`
+  const direction =
+    slideDirection === 'horizontal' || slideDirection === 'leftToRight'
+      ? 'left'
+      : 'right'
 
-  const resolveSlideDirection = () => {
-    switch (slideDirection) {
-      case 'horizontal':
-        return slideFromLeftToRight
-      case 'vertical':
-        return slideFromTopToBottom
-      case 'leftToRight':
-        return slideFromLeftToRight
-      case 'rightToLeft':
-        return slideFromRightToLeft
-      default:
-        return slideFromLeftToRight
-    }
-  }
+  const swipeHandler = direction === 'left' ? 'onSwipeLeft' : 'onSwipeRight'
 
   return (
     <>
@@ -179,47 +101,44 @@ const Drawer: StorefrontComponent<
       </div>
       <Portal>
         <Overlay visible={isMenuOpen} onClick={closeMenu} />
-
-        <Swipable
-          enabled={isMenuOpen}
-          element={menuRef && menuRef.current}
-          onSwipeLeft={
-            slideDirection === 'horizontal' || slideDirection === 'leftToRight'
-              ? closeMenu
-              : null
-          }
-          onSwipeRight={slideDirection === 'rightToLeft' ? closeMenu : null}
-          rubberBanding
-        >
-          <div
-            ref={menuRef}
-            className={`${handles.drawer} fixed top-0 ${
-              slideDirection === 'rightToLeft' ? 'right-0' : 'left-0'
-            } bottom-0 bg-base z-999 flex flex-column`}
+        <Suspense fallback={<React.Fragment />}>
+          <Swipable
+            {...{
+              [swipeHandler]: closeMenu,
+            }}
+            enabled={isMenuOpen}
+            position={isMenuOpen ? 'center' : direction}
+            allowOutsideDrag
+            className={`${handles.drawer} ${
+              direction === 'right' ? 'right-0' : 'left-0'
+            } fixed top-0 bottom-0 bg-base z-999 flex flex-column`}
             style={{
-              WebkitOverflowScrolling: 'touch',
-              overflowY: 'scroll',
               width: width || (isFullWidth ? '100%' : '85%'),
               maxWidth,
-              pointerEvents: isMenuOpen ? 'auto' : 'none',
-              transform: resolveSlideDirection(),
-              transition: isMenuTransitioning ? 'transform 300ms' : 'none',
               minWidth: 280,
+              pointerEvents: isMenuOpen ? 'auto' : 'none',
             }}
           >
-            <div className={`flex ${handles.closeIconContainer}`}>
-              <button
-                className={`pa4 pointer bg-transparent transparent bn pointer ${handles.closeIconButton}`}
-                onClick={closeMenu}
-              >
-                <IconClose size={30} type="line" />
-              </button>
+            <div
+              style={{
+                WebkitOverflowScrolling: 'touch',
+                overflowY: 'scroll',
+              }}
+            >
+              <div className={`flex ${handles.closeIconContainer}`}>
+                <button
+                  className={`${handles.closeIconButton} pa4 pointer bg-transparent transparent bn pointer`}
+                  onClick={closeMenu}
+                >
+                  <IconClose size={30} type="line" />
+                </button>
+              </div>
+              <div className={`${handles.childrenContainer} flex flex-grow-1`}>
+                {hasMenuBeenOpened && children}
+              </div>
             </div>
-            <div className={`${handles.childrenContainer} flex flex-grow-1`}>
-              {children}
-            </div>
-          </div>
-        </Swipable>
+          </Swipable>
+        </Suspense>
       </Portal>
     </>
   )