+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allows dragging from outside of the drawer.
+
+### Changed
+- Preserves momentum after swipe release.
+- Prevents rendering of content if the menu has not been opened.
+
+### Fixed
+- Issue where client rendering on the Portal component would be inconsistent with SSR.
+- "Accordeon effect" when opening the drawer, when the scrollbars are visible in the user's OS.
 
 ## [0.5.0] - 2019-12-27
 ### Changed